requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bleach==6.2.0
boto3==1.38.3
botocore==1.38.3
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.0.6
django-allauth==65.7.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==25.1
django-flexible-subscriptions==0.15.1
django-lifecycle==1.2.4
django-modeltranslation==0.19.13
django-parler==2.3
django-smart-selects==1.7.2
django-storages==1.14.6
django-timezone-field==7.1
django-unfold==0.40.0
django-unique-upload==0.2.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-api-logger==1.1.19
drf-spectacular==0.28.0
flake8==7.1.1
flower==2.0.1
h11==0.14.0
httptools==0.6.4
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
mccabe==0.7.0
msgpack==1.1.0
oauthlib==3.2.2
pillow==10.3.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
s3transfer==0.12.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.0
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.2
django-user-agents
django-cacheops
# django-seed
Faker
joblib
numpy
scikit-learn
tensorflow>=2.14,<2.18